requirements.txt
nltk
textblob
newspaper3k
Flask